Move translate test to `test_geometric_sources.py`
